1) Delete U-boot source of version 2018-07. 2) Add U-boot source of version 2015-10. 3) Modify 2015-10 vesion, and it can be compiled successfully by arm-linux-gcc(4.5.1).

+++ u-boot/u-boot-x4412/arch/arm/mach-bcm283x/mbox.c
@@ -1,6 +1,7 @@
-// SPDX-License-Identifier: GPL-2.0+
 /*
  * (C) Copyright 2012 Stephen Warren
+ *
+ * SPDX-License-Identifier:	GPL-2.0+
  */
 
 #include <common.h>
@@ -114,9 +115,7 @@ int bcm2835_mbox_call_prop(u32 chan, struct bcm2835_mbox_hdr *buffer)
 			   (unsigned long)((void *)buffer +
 			   roundup(buffer->buf_size, ARCH_DMA_MINALIGN)));
 
-	ret = bcm2835_mbox_call_raw(chan,
-				    phys_to_bus((unsigned long)buffer),
-				    &rbuffer);
+	ret = bcm2835_mbox_call_raw(chan, phys_to_bus((u32)buffer), &rbuffer);
 	if (ret)
 		return ret;
 
@@ -124,7 +123,7 @@ int bcm2835_mbox_call_prop(u32 chan, struct bcm2835_mbox_hdr *buffer)
 				(unsigned long)((void *)buffer +
 				roundup(buffer->buf_size, ARCH_DMA_MINALIGN)));
 
-	if (rbuffer != phys_to_bus((unsigned long)buffer)) {
+	if (rbuffer != phys_to_bus((u32)buffer)) {
 		printf("mbox: Response buffer mismatch\n");
 		return -1;
 	}
