+++ u-boot/u-boot-x4412/drivers/usb/gadget/udc/udc-core.c
@@ -1,4 +1,3 @@
-// SPDX-License-Identifier: GPL-2.0
 /**
  * udc-core.c - Core UDC Framework
  *
@@ -11,6 +10,8 @@
  *
  * commit 02e8c96627 : usb: gadget: udc: core: prepend udc_attach_driver with
  *		       usb_
+ *
+ * SPDX-License-Identifier:     GPL-2.0
  */
 
 #include <linux/compat.h>
@@ -64,7 +65,7 @@ void usb_gadget_unmap_request(struct usb_gadget *gadget,
 	if (req->length == 0)
 		return;
 
-	dma_unmap_single((void *)(uintptr_t)req->dma, req->length,
+	dma_unmap_single((void *)req->dma, req->length,
 			 is_in ? DMA_TO_DEVICE : DMA_FROM_DEVICE);
 }
 EXPORT_SYMBOL_GPL(usb_gadget_unmap_request);
